Modified ToEnumerable implementation to be slightly faster.
